apps/api/pyproject.toml
[project]
name = "oraculo-api"
version = "0.1.0"
description = "HTTP API for Oraculo Thorus"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0,<1.0.0",
    "uvicorn[standard]>=0.32.0,<1.0.0",
    "oraculo-ai",
    "langgraph-checkpoint-postgres>=3.0.0,<4.0.0",
    "psycopg[binary,pool]>=3.2.0",
    "python-dotenv>=1.0.0",
]

[tool.uv.sources]
oraculo-ai = { path = "../ai", editable = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/oraculo_api"]
